Finish native code cleanup
