+++ core/smart/detection/src/main/cpp/detection/matching_results.hpp
@@ -20,6 +20,7 @@
 
 #include <opencv2/core/types.hpp>
 #include <opencv2/imgproc/imgproc.hpp>
+
 #include "../types/scalable_roi.hpp"
 
 namespace smartautoclicker {
