+++ core/smart/detection/src/main/java/com/buzbuz/smartautoclicker/core/detection/ImageDetector.kt
@@ -17,9 +17,7 @@
 package com.buzbuz.smartautoclicker.core.detection
 
 import android.graphics.Bitmap
-import android.graphics.Point
 import android.graphics.Rect
-import androidx.annotation.Keep
 
 /**
  * Detects bitmaps within other bitmaps for conditions detection on the screen.
@@ -78,27 +76,3 @@ interface ImageDetector : AutoCloseable {
 const val DETECTION_QUALITY_MAX = 3216L
 /** The minimum detection quality for the algorithm. */
 const val DETECTION_QUALITY_MIN = 400L
-
-/**
- * The results of a condition detection.
- * @param isDetected true if the condition have been detected. false if not.
- * @param position contains the center of the detected condition in screen coordinates.
- * @param confidenceRate
- */
-data class DetectionResult(
-    var isDetected: Boolean = false,
-    val position: Point = Point(),
-    var confidenceRate: Double = 0.0
-) {
-
-    /**
-     * Set the results of the detection.
-     * Used by native code only.
-     */
-    @Keep
-    fun setResults(isDetected: Boolean, centerX: Int, centerY: Int, confidenceRate: Double) {
-        this.isDetected = isDetected
-        position.set(centerX, centerY)
-        this.confidenceRate = confidenceRate
-    }
-}
